refactor(carRacing) - Relocate cars in WinnerCarRacing.kt
